backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-dotenv==1.0.1

# Smartsheet API
smartsheet-python-sdk==3.0.3

# QR code generation
qrcode[pil]>=7.0
Pillow>=10.0

# PDF generation
reportlab>=4.0

# Google Cloud Storage
google-cloud-storage>=2.14.0

# Gmail API
google-auth>=2.25.0
google-api-python-client>=2.108.0

# HTTP requests
requests>=2.31.0
httpx>=0.27.0
